+++ src/components/base/b-slider/b-slider.ts
@@ -322,25 +322,48 @@ class bSlider extends iSliderProps implements iObserveDOM, iItems {
 	}
 
 	/**
+	 * Performs auto slide change.
+	 */
+	protected performAutoSlide(): void {
+		void this.removeMod('swipe');
+		this.moveSlide(1);
+		this.syncState();
+		void this.removeMod('swipe');
+	}
+
+	/**
+	 * Plays auto slide changes at a given interval.
+	 */
+	protected playAutoSlide(): void {
+		this.async.setInterval(
+			() => this.performAutoSlide(),
+			this.autoSlideInterval,
+			{label: $$.autoSlide}
+		);
+	}
+
+	/**
 	 * Resumes auto slide moves by setting the corresponding interval.
+	 * @param firstInterval - an interval (in ms) before first auto slide change.
 	 */
 	@hook('mounted')
 	@watch('autoSlideInterval')
-	protected resumeAutoSlide(): void {
-		if (this.isSlideMode && Number.isPositive(this.autoSlideInterval)) {
-			this.async.setInterval(
-				() => {
-					void this.removeMod('swipe');
-					this.moveSlide(1);
-					this.syncState();
-					void this.removeMod('swipe');
-				},
-				this.autoSlideInterval,
-				{label: $$.autoSlide}
-			);
-		} else {
+	protected resumeAutoSlide(firstInterval: number = this.autoSlideInterval): void {
+		if (!this.isSlideMode || !Number.isPositive(this.autoSlideInterval)) {
 			this.pauseAutoSlide();
+			return;
 		}
+
+		firstInterval = Math.max(this.autoSlideInterval, firstInterval);
+
+		this.async.setTimeout(
+			() => {
+				this.performAutoSlide();
+				this.playAutoSlide();
+			},
+			firstInterval,
+			{label: $$.autoSlideFirst}
+		);
 	}
 
 	/**
@@ -628,7 +651,7 @@ class bSlider extends iSliderProps implements iObserveDOM, iItems {
 		this.emit('swipeEnd', dir, isSwiped);
 		this.isTolerancePassed = false;
 		this.swiping = false;
-		this.resumeAutoSlide();
+		this.resumeAutoSlide(this.autoSlidePostGestureDelay);
 	}
 }