+++ src/components/base/b-slider/props.ts
@@ -78,6 +78,12 @@ export default abstract class iSliderProps extends iData {
 	readonly alignFirstToStart: boolean = true;
 
 	/**
+	 * If true, the last slide will be aligned to the end position (the right bound).
+	 */
+	@prop(Boolean)
+	readonly alignLastToEnd: boolean = true;
+
+	/**
 	 * How much does the shift along the X-axis corresponds to a finger movement
 	 */
 	@prop({type: Number, validator: (v) => Number.isPositiveBetweenZeroAndOne(v)})
@@ -112,4 +118,19 @@ export default abstract class iSliderProps extends iData {
 	 */
 	@prop({type: Number, validator: (v) => Number.isNatural(v)})
 	readonly swipeToleranceY: number = 50;
+
+	/**
+	 * The interval (in milliseconds) between auto slide moves.
+	 * If it is set to `0`, it means that there will be no auto slide moves.
+	 */
+	@prop({type: Number, validator: (v) => Number.isNonNegative(v)})
+	readonly autoSlideInterval: number = 0;
+
+	/**
+	 * The delay (in milliseconds) between the last user gesture and the first automatic slide movement.
+	 * It will be capped at the maximum value between `autoSlideInterval` and `autoSlidePostGestureDelay`,
+	 * and will be used as a timeout for the first automatic slide movement after a user gesture.
+	 */
+	@prop({type: Number, validator: (v) => Number.isNonNegative(v)})
+	readonly autoSlidePostGestureDelay: number = 0;
 }