+++ src/core/prelude/test-env/gestures.ts
@@ -26,6 +26,58 @@ export default class Gestures {
 	steps: Array<Required<TouchGesturePoint>> = [];
 
 	/**
+	 * Dispatches a touch event.
+	 * It is intended for use in cases where the standard functionality of Gestures is not suitable for
+	 * solving a specific problem:
+	 *
+	 * 1. If you need to pass several points in one event, you can pass an array of coordinates as the second parameter.
+	 * 2. If only the emission of a specific event (touchstart, touchmove, touchend) is required,
+	 *    you can fill the last two parameters with the corresponding elements or selectors.
+	 *
+	 * 3. If you want to emit a touch event over the entire document,
+	 *    and not over a specific element, you can omit the last two parameters.
+	 *
+	 * @param eventType - the type of the event
+	 * @param touchPoints - a point or an array of points for touches
+	 * @param [targetEl] - the target element, defaults to `document.documentElement`
+	 * @param [dispatchEl] - the element from which the event is dispatched,
+	 *   defaults to `document.elementFromPoint(<first point>)`
+	 */
+	static dispatchTouchEvent(
+		eventType: 'touchstart' | 'touchmove' | 'touchend',
+		touchPoints: CanArray<{x: number; y: number}>,
+		targetEl: CanNull<Element> = null,
+		dispatchEl: CanNull<Element> = null
+	): void {
+		if (!Object.isArray(touchPoints)) {
+			touchPoints = [touchPoints];
+		}
+
+		if (targetEl == null) {
+			targetEl = document.documentElement;
+		}
+
+		if (dispatchEl == null) {
+			const {x, y} = touchPoints[0];
+			dispatchEl = document.elementFromPoint(x, y);
+		}
+
+		const event = new TouchEvent(eventType, {
+			bubbles: true,
+			cancelable: true,
+			composed: true,
+			touches: touchPoints.map<Touch>(({x: clientX, y: clientY}, identifier) => new Touch({
+				identifier,
+				clientX,
+				clientY,
+				target: targetEl!
+			}))
+		});
+
+		dispatchEl!.dispatchEvent(event);
+	}
+
+	/**
 	 * @param opts
 	 */
 	constructor(opts: TouchGesturesCreateOptions) {
@@ -135,26 +187,15 @@ export default class Gestures {
 			{dispatchEl, targetEl, x, y} = step;
 
 		const
-			resolvedDispatchEl = dispatchEl instanceof Element ? dispatchEl : document.querySelector(dispatchEl),
-			resolvedTargetEl = targetEl instanceof Element ? targetEl : document.querySelector(targetEl);
-
-		const touchEvent = new TouchEvent(type, {
-			touches: [
-				new Touch({
-					identifier: Math.random(),
-					target: resolvedTargetEl!,
-					clientX: x,
-					clientY: y
-				})
-			]
-		});
+			resolvedDispatchEl = this.resolveElement(dispatchEl),
+			resolvedTargetEl = this.resolveElement(targetEl);
 
 		Object.assign(this.cursor.style, {
 			left: x.px,
 			top: y.px
 		});
 
-		resolvedDispatchEl?.dispatchEvent(touchEvent);
+		Gestures.dispatchTouchEvent(type, {x, y}, resolvedTargetEl, resolvedDispatchEl);
 	}
 
 	/**
@@ -182,6 +223,21 @@ export default class Gestures {
 			this.steps.push(newPoint);
 		});
 	}
+
+	/**
+	 * Returns a DOM node if the passed element is a DOM node or
+	 * performs a `querySelector` to find a DOM node based on the passed selector
+	 *
+	 * @param element - an element to resolve
+	 */
+	protected resolveElement(element: Required<TouchGesturesCreateOptions>['targetEl']): CanNull<Element> {
+		if (element instanceof Element) {
+			return element;
+		}
+
+		return document.querySelector(element);
+	}
+
 }
 
 globalThis._Gestures = Gestures;