+++ src/core/prelude/webpack.ts
@@ -6,11 +6,13 @@
  * https://github.com/V4Fire/Client/blob/master/LICENSE
  */
 
-/* eslint-disable camelcase,no-new-func */
+/* eslint-disable camelcase */
+
+import global from 'core/shims/global';
 
 try {
 	const
-		ctx = Function('return this')();
+		ctx = global;
 
 	__webpack_nonce__ = ctx[CSP_NONCE_STORE];
 	__webpack_public_path__ = ctx.PUBLIC_PATH ?? PUBLIC_PATH;